destalles de formulario

+++ app/Http/Controllers/EmpleadoController.php
@@ -51,7 +51,8 @@ class EmpleadoController extends Controller
         }
 
         Empleado::insert($datosEmpleado);
-        return response()->json($datosEmpleado);
+        //return response()->json($datosEmpleado);
+        return redirect('empleado')->with('mensaje','Empleado agregado con exito');
     }
 
     /**
@@ -117,8 +118,18 @@ class EmpleadoController extends Controller
     public function destroy( $id)
     {
         //
-        Empleado::destroy($id);
-        return redirect ('empleado');
+        $empleado=Empleado::findOrFail($id);
+
+        if(Storage::delete('public/'.$empleado->foto)){
+            Empleado::destroy($id);
+        }
+
+
+
+
+
+       
+        return redirect ('empleado')->with('mensaje','Empleado Borrado');
 
     }
 }
